use a structured data format to publish waivers and waiver requests
Rather than using the Federal Register to publish notices of waiver requests (nb. this, as written, is not the requests themselves), let's have OMB publish the waivers actually granted in a table, with the full waiver request where it is sensible to do so, or else an explanation why the request isn't posted.