avoid implying that any contract warrants IP protections for the contractor
This paragraph comes from the M-13-13 memo, which tried to direct agencies to make open data without directing them to always make open data because sometimes it's OK to give the contractor the IP, trading IP rights for dollars on the contract, effectively.

OK. Fine. But let's be clearer about it. As written, including in M-13-13, the text sets up a tension between open data and "proprietary interests." That's not the right tension. It should be between open data and "encourag[ing] qualified contractors to participate." The Federal government shouldn't care about proprietary interests for their own sake.

I've made a minimal change that I think gets at this ---- simply splitting the sentence into two:

> For all instances where new Federal information creation or collection does not fall squarely within the public domain as U.S. government work, agencies shall include appropriate provisions in contracts to meet objectives of open data. [SENTENCE BREAK HERE] Recognizing that contractors may have proprietary interests in such information, protection of such information may be necessary to encourage qualified contractors to participate in and apply innovative concepts to government programs.