further clarifications
"accessible" is defined in Definitions to mean section 508 compliance. This differs from how the term is used in the open data community, where it compasses variously: available, usable for a wide range of purposes, published in bulk, documented, etc. This clarifies that accessible is used here as it is defined in Definitions, and I'm explicitly adding "widest possible range of purposes" and "in bulk" to cover the other aspects of the meaning. (Documented is already covered by "described.")